+++ framework/ui/src/components/index.ts
@@ -7,6 +7,7 @@ export { default as VcCol } from "./atoms/vc-col/vc-col.vue";
 export { default as VcContainer } from "./atoms/vc-container/vc-container.vue";
 export { default as VcHint } from "./atoms/vc-hint/vc-hint.vue";
 export { default as VcIcon } from "./atoms/vc-icon/vc-icon.vue";
+export { default as VcInfoRow } from "./atoms/vc-info-row/vc-info-row.vue";
 export { default as VcImage } from "./atoms/vc-image/vc-image.vue";
 export { default as VcLabel } from "./atoms/vc-label/vc-label.vue";
 export { default as VcLink } from "./atoms/vc-link/vc-link.vue";
@@ -14,24 +15,23 @@ export { default as VcLoading } from "./atoms/vc-loading/vc-loading.vue";
 export { default as VcProgress } from "./atoms/vc-progress/vc-progress.vue";
 export { default as VcRow } from "./atoms/vc-row/vc-row.vue";
 export { default as VcStatus } from "./atoms/vc-status/vc-status.vue";
-export { default as VcSwitch } from "./atoms/vc-switch/vc-switch.vue";
-export { default as VcTextarea } from "./atoms/vc-textarea/vc-textarea.vue";
-export { default as VcWidget } from "./atoms/vc-widget/vc-widget.vue";
-export { default as VcInfoRow } from "./atoms/vc-info-row/vc-info-row.vue";
 export { default as VcStatusIcon } from "./atoms/vc-status-icon/vc-status-icon.vue";
-export { default as VcRating } from "./atoms/vc-rating/vc-rating.vue";
+export { default as VcSwitch } from "./atoms/vc-switch/vc-switch.vue";
+export { default as VcWidget } from "./atoms/vc-widget/vc-widget.vue";
 
 /* Molecules */
 export { default as VcBreadcrumbs } from "./molecules/vc-breadcrumbs/vc-breadcrumbs.vue";
+export { default as VcEditor } from "./molecules/vc-editor/vc-editor.vue";
 export { default as VcForm } from "./molecules/vc-form/vc-form.vue";
+export { default as VcFileUpload } from "./molecules/vc-file-upload/vc-file-upload.vue";
 export { default as VcInput } from "./molecules/vc-input/vc-input.vue";
 export { default as VcMultivalue } from "./molecules/vc-multivalue/vc-multivalue.vue";
 export { default as VcNotification } from "./molecules/vc-notification/vc-notification.vue";
 export { default as VcPagination } from "./molecules/vc-pagination/vc-pagination.vue";
+export { default as VcRating } from "./molecules/vc-rating/vc-rating.vue";
 export { default as VcSelect } from "./molecules/vc-select/vc-select.vue";
-export { default as VcFileUpload } from "./molecules/vc-file-upload/vc-file-upload.vue";
 export { default as VcSlider } from "./molecules/vc-slider/vc-slider.vue";
-export { default as VcEditor } from "./molecules/vc-editor/vc-editor.vue";
+export { default as VcTextarea } from "./molecules/vc-textarea/vc-textarea.vue";
 
 /* Organisms */
 export { default as VcApp } from "./organisms/vc-app/vc-app.vue";
